+++ ChangeLog.md
@@ -6,6 +6,7 @@ ChangeLog
 
 * Changed: iTip broker now sets RSVP status to false when replies are received.
 * #118: iTip Message now has a `getScheduleStatus()` method.
+* #119: Support for detecting 'significant changes'.
 
 
 3.3.1 (2014-08-18)